Move DeckFactory code into Deck, so that Decks can only be made through static factory methods
