+++ src/test/java/DeckTest.java
@@ -1,17 +1,21 @@
+import dictionary.Word;
+import fixtures.WordFixtures;
 import hanziToAnki.Deck;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import java.util.List;
+
+import java.util.Set;
 
 class DeckTest {
 
     @Test
     void readsLinesFromFile() {
-        Deck deck = new Deck();
-        deck.addLine("This is a line");
-        deck.addLine("This is another line, aEIORxfi;lhmgzfAD," + "another");
+        Word word1 = WordFixtures.aWord();
+        Word word2 = WordFixtures.aWord();
 
-        List<String> lines = deck.getLines();
-        Assertions.assertEquals(2, lines.size());
+        var words = Set.of(word1, word2);
+        var deck = Deck.generateDeck(words);
+
+        Assertions.assertEquals(2, deck.getLines().size());
     }
 }
