+++ src/test/java/fixtures/WordFixtures.java
@@ -1,0 +1,16 @@
+package fixtures;
+
+import dictionary.Word;
+
+public class WordFixtures {
+
+    public static Word aWord() {
+        return new Word(
+        "traditional",
+                "simplified",
+                "pinyin",
+                "pinyinTones",
+                "definition"
+        );
+    }
+}
